feat(todo): Move to todo package, simplify, create Deadline/Event
