+++ src/main/java/cloud/main/todo/Todo.java
@@ -1,4 +1,4 @@
-package cloud.main;
+package cloud.main.todo;
 
 
 
@@ -6,16 +6,12 @@ package cloud.main;
  * Represents a TODO.
  */
 public class Todo {
+	private boolean isComplete = false;
+
 	private String description;
-	private boolean isComplete;
 
-	public Todo(String _description) {
-		this(_description, false);
-	}
-
-    public Todo(String _description, boolean _isComplete) {
+    public Todo(String _description) {
 		this.description = _description;
-		this.isComplete = _isComplete;
     }
 
 	public String getDescription() {
